refactor: :art: urlの変更により、フォルダ名の変更
